test(fish-s2pro): consolidate forward regression test
